convert dtype to float

+++ s3fd.py
@@ -152,5 +152,5 @@ class s3fd_original(nn.Module):
         print(cls1.size())
         #return [cls1,reg1,cls2,reg2,cls3,reg3,cls4,reg4,cls5,reg5,cls6,reg6]
         #return [cls1,cls2,cls3,cls4,cls5,cls6]
-        return torch.from_numpy(array([cls2,cls3,cls4,cls5]))
+        return torch.from_numpy(array([cls2,cls3,cls4,cls5]).astype('float'))
         #return [cls2,gen2,cls3,gen3,cls4,gen4,cls5,gen5]
